+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Builds osx start/stop notification strings from launchd.plist values
 - Updated user name validation to match Web Control
+- Added --force option to config user to allow legacy names
 
 ---